change int -> size_t, ft_calloc() -> malloc()

+++ inc/tokenize_utils.h
@@ -6,7 +6,7 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/15 17:44:13 by pnguyen-          #+#    #+#             */
-/*   Updated: 2024/03/28 09:07:47 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2024/03/28 10:22:03 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef enum e_action
 typedef struct s_line_part
 {
 	char const	*line;
-	int			index;
+	size_t		index;
 	int			mode;
 	t_token		*token;
 	t_list		*last_node;
